+++ app/src/main/java/internal/http/exception/MtraceHttpConnectException.java
@@ -1,0 +1,10 @@
+package internal.http.exception;
+
+
+import javax.validation.constraints.NotNull;
+
+public final class MtraceHttpConnectException extends RuntimeException {
+    public MtraceHttpConnectException(@NotNull String errorMessage, @NotNull Exception originalException) {
+        super(errorMessage, originalException);
+    }
+}
