+++ app/src/main/java/internal/http/model/MtraceHttpMethod.java
@@ -1,4 +1,4 @@
-package util.http.model;
+package internal.http.model;
 
 public enum MtraceHttpMethod {
    GET;
