+++ app/src/main/java/internal/http/model/MtraceHttpRequest.java
@@ -1,4 +1,4 @@
-package util.http.model;
+package internal.http.model;
 
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
