--- app/src/main/java/util/http/MtraceHttpClient.java
@@ -1,8 +1,0 @@
-package util.http;
-
-import util.http.model.MtraceHttpRequest;
-import util.http.model.MtraceHttpResponse;
-
-public interface MtraceHttpClient {
-   MtraceHttpResponse get(MtraceHttpRequest request) throws Exception;
-}
