--- app/src/main/java/util/http/exception/MtraceHttpConnectException.java
@@ -1,10 +1,0 @@
-package util.http.exception;
-
-
-import javax.validation.constraints.NotNull;
-
-public final class MtraceHttpConnectException extends RuntimeException {
-   public MtraceHttpConnectException(@NotNull String errorMessage, @NotNull Exception originalException) {
-      super(errorMessage, originalException);
-   }
-}
